forager_embedding_server/containers/resizer/requirements.txt
aiohttp[speedups]
gcloud-aio-storage
Pillow
sanic
uvicorn
sanic_compress
